move loadScript() and related functions into new script module
also move tests to karma-typescript
