+++ test/esri-loader.spec.js
@@ -27,224 +27,6 @@ function removeRequire() {
 
 describe('esri-loader', function () {
   const jaspi3xUrl = 'base/test/mocks/jsapi3x.js';
-  describe('when has not yet been loaded', function () {
-    beforeEach(function() {
-      removeRequire();
-      removeScript();
-    });
-    it('isLoaded should be false', function () {
-      expect(esriLoader.isLoaded()).toBeFalsy();
-    });
-  });
-
-  describe('when loading the script', function () {
-    describe('with defaults', function () {
-      var scriptEl;
-      beforeAll(function (done) {
-        spyOn(document.body, 'appendChild').and.callFake(function (el) {
-          // trigger the onload event listeners
-          el.dispatchEvent(new Event('load'));
-        });
-        spyOn(document.head, 'appendChild');
-        esriLoader.loadScript()
-        .then((script) => {
-          // hold onto script element for assertions below
-          scriptEl = script;
-          done();
-        });
-      });
-      it('should default to latest version', function () {
-        expect(scriptEl.src).toEqual('https://js.arcgis.com/4.10/');
-      });
-      it('should not have set dojoConfig', function () {
-        expect(window.dojoConfig).not.toBeDefined();
-      });
-      it('should not have called loadCss', function () {
-        expect(document.head.appendChild.calls.any()).toBeFalsy();
-      });
-    });
-    describe('with different API version', function () {
-      var scriptEl;
-      beforeAll(function (done) {
-        spyOn(document.body, 'appendChild').and.callFake(function (el) {
-          // trigger the onload event listeners
-          el.dispatchEvent(new Event('load'));
-        });
-        esriLoader.loadScript({
-          url: 'https://js.arcgis.com/3.27'
-        })
-        .then((script) => {
-          // hold onto script element for assertions below
-          scriptEl = script;
-          done();
-        });
-      });
-      it('should load different version', function () {
-        expect(scriptEl.src).toEqual('https://js.arcgis.com/3.27');
-      });
-      it('should not have set dojoConfig', function () {
-        expect(window.dojoConfig).not.toBeDefined();
-      });
-    });
-    describe('with css option', function () {
-      var cssUrl = 'https://js.arcgis.com/4.10/esri/css/main.css';
-      beforeAll(function (done) {
-        spyOn(document.body, 'appendChild').and.callFake(function (el) {
-          // trigger the onload event listeners
-          el.dispatchEvent(new Event('load'));
-        });
-        spyOn(document.head, 'appendChild').and.stub();
-        esriLoader.loadScript({
-          css: cssUrl
-        })
-        .then((script) => {
-          done();
-        });
-      });
-      it('should have called loadCss with the url', function () {
-        expect(document.head.appendChild.calls.argsFor(0)[0].href).toEqual(cssUrl);
-      });
-    });
-    describe('with dojoConfig option', function () {
-      var dojoConfig = {
-        async: true,
-        packages: [
-          {
-            location: 'path/to/somelib',
-            name: 'somelib'
-          }
-        ]
-      };
-      beforeAll(function (done) {
-        spyOn(document.body, 'appendChild').and.callFake(function (el) {
-          // trigger the onload event listeners
-          el.dispatchEvent(new Event('load'));
-        });
-        esriLoader.loadScript({
-          dojoConfig: dojoConfig
-        })
-        .then((script) => {
-          done();
-        });
-      });
-      it('should have set global dojoConfig', function () {
-        expect(window.dojoConfig).toEqual(dojoConfig);
-      });
-      afterAll(function() {
-        window.dojoConfig = undefined;
-      });
-    });
-    describe('when already loaded by some other means', function () {
-      beforeAll(function () {
-        stubRequire();
-      });
-      it('should reject', function (done) {
-        esriLoader.loadScript({
-          url: jaspi3xUrl
-        })
-        .then(script => {
-          done.fail('call to loadScript should have failed');
-        })
-        .catch(err => {
-          expect(err.message).toEqual(`The ArcGIS API for JavaScript is already loaded.`);
-          done();
-        });
-      });
-      afterAll(function () {
-        // clean up
-        removeRequire();
-      });
-    });
-    describe('when loading an invalid url', function () {
-      it('should pass an error to the callback', function (done) {
-        esriLoader.loadScript({
-          url: 'not a valid url'
-        })
-        .then(script => {
-          done.fail('call to loadScript should have failed');
-        })
-        .catch(err => {
-          expect(err.message.indexOf('There was an error attempting to load')).toEqual(0);
-          done();
-        });
-      });
-      afterAll(function () {
-        // clean up
-        removeScript();
-      });
-    });
-    describe('when called twice', function () {
-      describe('when loading the same script', function () {
-        it('should resolve the script if it is already loaded', function (done) {
-          esriLoader.loadScript({
-            url: jaspi3xUrl
-          })
-          .then(firstScript => {
-            // try loading the same script after the first one has already loaded
-            esriLoader.loadScript({
-              url: jaspi3xUrl
-            })
-            .then(script => {
-              expect(script.getAttribute('src')).toEqual(jaspi3xUrl);
-              done();
-            })
-            .catch(err => {
-              done.fail('second call to loadScript should not have failed with: ' + err);
-            });
-          })
-          .catch(() => {
-            done.fail('first call to loadScript should not have failed');
-          });
-        });
-        it('should resolve an unloaded script once it loads', function (done) {
-          esriLoader.loadScript({
-            url: jaspi3xUrl
-          })
-          .catch(() => {
-            done.fail('first call to loadScript should not have failed');
-          });
-          // try loading the same script again
-          esriLoader.loadScript({
-            url: jaspi3xUrl
-          })
-          .then(script => {
-            expect(script.getAttribute('src')).toEqual(jaspi3xUrl);
-            done();
-          })
-          .catch(err => {
-            done.fail('second call to loadScript should not have failed with: ' + err);
-          });
-        });
-      });
-      describe('when loading different scripts', function () {
-        it('should reject', function (done) {
-          esriLoader.loadScript({
-            url: jaspi3xUrl
-          })
-          .catch(() => {
-            done.fail('first call to loadScript should not have failed');
-          });
-          // try loading a different script
-          esriLoader.loadScript({
-            url: 'base/test/mocks/jsapi4x.js'
-          })
-          .then(script => {
-            done.fail('second call to loadScript should have failed');
-          })
-          .catch(err => {
-            expect(err.message).toEqual(`The ArcGIS API for JavaScript is already loaded (${jaspi3xUrl}).`);
-            done();
-          });
-        });
-      });
-      afterEach(function () {
-        // clean up
-        removeRequire();
-        removeScript();
-      });
-    });
-  });
-
   describe('when loading modules', function () {
     var expectedModuleNames = ['esri/map', 'esri/layers/VectorTileLayer'];
     describe('when script has been loaded', function() {
