+++ test/mocks/jsapi4x.js
@@ -1,3 +1,3 @@
-// this is defined in spec
 console.log('4.x');
+// NOTE: this is defined in /src/script.test.ts
 stubRequire();
